+++ packages/design-system/components/web/inputs/MarkdownEditorField/types.ts
@@ -4,6 +4,7 @@ import type {
   FC,
   KeyboardEventHandler,
   ReactNode,
+  Ref,
 } from 'react'
 
 import type { FormControl } from '@baseapp-frontend/utils'
@@ -31,6 +32,7 @@ export interface ToolbarConfig {
 
 export type EditorContainerProps = BoxProps & {
   minHeight?: number | string
+  maxHeight?: number | string
   hasBorder?: boolean
   hasLabel?: boolean
   error?: boolean
@@ -51,13 +53,17 @@ export type MarkdownEditorFieldCoreProps = Omit<MDXEditorProps, 'markdown'> & {
   onKeyDown?: KeyboardEventHandler<HTMLTextAreaElement>
   placeholder?: string
   onChange?: (value: string) => void
+  inputRef?: Ref<HTMLInputElement>
   toolbarConfig?: Partial<ToolbarConfig>
   showDiffSourceToggle?: boolean
   showUndoRedo?: boolean
   minHeight?: number | string
+  maxHeight?: number | string
   hasBorder?: boolean
   label?: ReactNode
+  labelBackgroundColor?: string
   helperText?: ReactNode
+  showHelperText?: boolean
   error?: boolean
   Toolbar?: FC<ToolbarProps>
   ToolbarProps?: Partial<ToolbarProps>